requirements.txt
msgpack
chromadb
openai
python-dotenv
